+++ requests/name.go
@@ -1,12 +1,12 @@
-package main
+package requests
 
 const nameUrl = "https://names.mcquay.me/api/v0/"
 
-type name struct {
+type Name struct {
 	FirstName string `json:"first_name"`
 	LastName  string `json:"last_name"`
 }
 
-func (n *name) GetUrl() string {
+func (n *Name) GetUrl() string {
 	return nameUrl
 }
